requirements.txt
pytest
allure-pytest
selene>=2.0.0rc9
Appium-Python-Client==4.2.1
selenium==4.21.0
